--- src/main/kotlin/de/jan/jdaktx/CommandHandler/commands/KSubCommands.kt
@@ -1,15 +1,0 @@
-package de.jan.jdaktx.CommandHandler.commands
-
-import net.dv8tion.jda.api.interactions.commands.build.SubcommandData
-
-class KSubCommands {
-
-    internal val commands = mutableListOf<SubcommandData>()
-
-    fun subCommand(name: String = "", description: String = "", cmd: KSubCommand.() -> Unit) {
-        val subCommand = KSubCommand(name, description)
-        subCommand.cmd()
-        commands.add(subCommand.subCommand)
-    }
-
-}
