--- src/main/kotlin/de/jan/jdaktx/KClasses/KEmbeds/KEmbedField.kt
@@ -1,9 +1,0 @@
-package de.jan.jdaktx.KClasses.KEmbeds
-
-class KEmbedField {
-
-    var name: String = ""
-    var value: String = ""
-    var inline: Boolean = false
-
-}
