+++ src/main/kotlin/de/jan/jdaktx/classes/Permissions.kt
@@ -1,0 +1,18 @@
+package de.jan.jdaktx.classes
+
+import net.dv8tion.jda.api.Permission
+
+class Permissions {
+
+    val allow = mutableListOf<Permission>()
+    val deny = mutableListOf<Permission>()
+
+    operator fun Permission.unaryPlus() {
+        allow.add(this)
+    }
+
+    operator fun Permission.unaryMinus() {
+        deny.add(this)
+    }
+
+}
