+++ src/main/kotlin/de/jan/jdaktx/classes/embeds/KEmbedField.kt
@@ -1,0 +1,9 @@
+package de.jan.jdaktx.classes.embeds
+
+class KEmbedField {
+
+    var name: String = ""
+    var value: String = ""
+    var inline: Boolean = false
+
+}
